fixed the inbox not removed for Http.ERROR

+++ backend/src/main/java/com/taskhive/backend/controller/ProjectInviteController.java
@@ -34,27 +34,35 @@ public class ProjectInviteController {
     @Autowired
     private InboxService inboxService;
 
+    /**
+     * NOTE: when this route is invoked and if the inbox exists in the databaes for the current user, the Inbox is deleted for the user, regardless whether the response is ERROR or OK
+     */
+
     @GetMapping("/acceptProjectInvite")
     public ResponseEntity<Response> acceptProjectInvite(@RequestHeader String pid, @RequestHeader String inboxId, Authentication authentication) {
         AppUser user = appUserService.loadUserByUsername(authentication.getName());
 
-        //1) check if the user has already joined the project
-        List<Project> joinedProjects = user.getJoinedProjects();
-        for (Project project : joinedProjects) {
-            if (Integer.toString(project.getPid()).equals(pid)) {
-                return new ResponseEntity<>(new Response(ResponseMessage.PROJECT_ALREADY_JOINED), HttpStatus.NOT_ACCEPTABLE);
-            }
-        }
-
-        //2) check if the inbox exists in the database for the current user
+        //1) check if the inbox exists in the database for the current user
         Inbox inbox = inboxService.getInboxById(Integer.parseInt(inboxId));
         if (inbox == null || inbox.getUser() == null || inbox.getUser().getUid() != user.getUid()) {
             return new ResponseEntity<>(new Response(ResponseMessage.INBOX_DOES_NOT_EXIST), HttpStatus.NOT_ACCEPTABLE);
         }
 
-        //2) check if the invitation exists and is a valid invitation
-        //3) add the project into the user's joinedProject and save user
-        //4) delete the inbox from user
+
+        //2) check if the user has already joined the project
+        List<Project> joinedProjects = user.getJoinedProjects();
+        for (Project project : joinedProjects) {
+            if (Integer.toString(project.getPid()).equals(pid)) {
+                inboxService.deleteInbox(inbox);
+                return new ResponseEntity<>(new Response(ResponseMessage.PROJECT_ALREADY_JOINED), HttpStatus.NOT_ACCEPTABLE);
+            }
+        }
+
+
+        //3) check if the invitation exists and is a valid invitation
+        //4) test the whether the targetProject is null and targetProject's status is COMPLETED
+        //then, add the project into the user's joinedProject and save user
+        //then, delete the inbox from user
         if (inbox.getPid() == Integer.parseInt(pid)
                 && inbox.getTitle() == InboxInviteTitle.INVITATION) {
 
@@ -65,7 +73,7 @@ public class ProjectInviteController {
              * add the project to joinedProjects and since cascade = persist, the project data is automatically saved too
              * */
 
-            if (targetProject.getStatus() != ProjectStatus.COMPLETED) {
+            if (targetProject != null && targetProject.getStatus() != ProjectStatus.COMPLETED) {
                 user.getJoinedProjects().add(targetProject);
                 user.getInboxes().remove(inbox);
                 appUserService.update(user);
@@ -80,8 +88,10 @@ public class ProjectInviteController {
 
                 return new ResponseEntity<>(new Response(ResponseMessage.INBOX_INVITATION_ACCEPTED), HttpStatus.OK);
             }
-            return new ResponseEntity<>(new Response(ResponseMessage.INBOX_INVITATION_EXPIRED), HttpStatus.NOT_ACCEPTABLE);
+            inboxService.deleteInbox(inbox);
+            return new ResponseEntity<>(new Response(ResponseMessage.INBOX_INVITATION_INVALID), HttpStatus.NOT_ACCEPTABLE);
         }
+        inboxService.deleteInbox(inbox);
         return new ResponseEntity<>(new Response(ResponseMessage.ERROR_OCCURED), HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
@@ -90,27 +100,37 @@ public class ProjectInviteController {
     public ResponseEntity<Response> rejectProjectInvite(@RequestHeader String pid, @RequestHeader String inboxId, Authentication authentication) {
         AppUser user = appUserService.loadUserByUsername(authentication.getName());
 
-        //1) check if the user has already joined the project
-        List<Project> joinedProjects = user.getJoinedProjects();
-        for (Project project : joinedProjects) {
-            if (Integer.toString(project.getPid()).equals(pid)) {
-                return new ResponseEntity<>(new Response(ResponseMessage.PROJECT_ALREADY_JOINED), HttpStatus.NOT_ACCEPTABLE);
-            }
-        }
-
-        //2, 3) check if the inbox exists in the database for the current user
+        //1) check if the inbox exists in the database and the inbox is for the current user
         Inbox inbox = inboxService.getInboxById(Integer.parseInt(inboxId));
         if (inbox == null || inbox.getUser() == null || inbox.getUser().getUid() != user.getUid()) {
             return new ResponseEntity<>(new Response(ResponseMessage.INBOX_DOES_NOT_EXIST), HttpStatus.NOT_ACCEPTABLE);
         }
 
-        //4 check if the pid in the inbox is the same as in the header
-        //5 check if the inbox is an invitation
+        //2) check if the user has already joined the project
+        List<Project> joinedProjects = user.getJoinedProjects();
+        for (Project project : joinedProjects) {
+            if (Integer.toString(project.getPid()).equals(pid)) {
+                inboxService.deleteInbox(inbox);
+                return new ResponseEntity<>(new Response(ResponseMessage.PROJECT_ALREADY_JOINED), HttpStatus.NOT_ACCEPTABLE);
+            }
+        }
+
+
+        //3 check if the pid in the inbox is the same as in the header and check if the inbox is an invitation
         if (inbox.getTitle() == InboxInviteTitle.INVITATION && inbox.getPid() == Integer.parseInt(pid)) {
-            inboxService.deleteInbox(inbox);
 
             //create an inbox for the initiator
             Project targetProject = projectService.loadProjectByPid(Integer.parseInt(pid));
+
+            //4 check whether the targetProject exists in the databaase
+            if (targetProject == null) {
+                inboxService.deleteInbox(inbox);
+                return new ResponseEntity<>(new Response(ResponseMessage.INBOX_INVITATION_INVALID), HttpStatus.NOT_ACCEPTABLE);
+            }
+
+            //we could have also detached the inbox from the user and then saved and orphanRemoval would have removed the record
+            inboxService.deleteInbox(inbox);
+
             AppUser initiator = appUserService.loadUserByUsername(inbox.getInitiator());
             Inbox inboxForInitiator = Inbox.builder().user
                     (initiator).title(InboxInviteTitle.PROJECT_REJECTED).initiator(user.getUsername()).projectName(targetProject.getProjectName()).build();
@@ -121,6 +141,7 @@ public class ProjectInviteController {
 
             return new ResponseEntity<>(new Response(ResponseMessage.INBOX_INVITATION_DELETED), HttpStatus.OK);
         }
+        inboxService.deleteInbox(inbox);
         return new ResponseEntity<>(new Response(ResponseMessage.ERROR_OCCURED), HttpStatus.INTERNAL_SERVER_ERROR);
     }
 }
